Revert "PYTHON-5930 Move TestClientSSL to a shared, non-mirrored test file"
This reverts commit 9e55d86e6971779f78e736043a423f6eeaffc002.
